// src/types.ts
export type JsonValue = JsonPrimitive | JsonObject | JsonValue[];
export interface JsonObject {
  [key: string]: JsonValue;
}

export type AgentInteractionMode = "point_to_point" | "group" | "hybrid";
export type AgentRole = "requester" | "service";
export type MessageArtifact = "work_communication" | "work_result";
export type TaskState =
  | "accepted"
  | "rejected"
  | "completed"
  | "failed"
  | "cancelled"
  | "in_progress"
  | "progress";

export interface AgentIdentityCodeParts {
  oid: "1.2.156.3088";
  version: "1";
  registrationServiceProvider: string;
  registrationRequester: string;
  ontologySerial: string;
  instanceSerial: string;
}

export type AgentIdentityCode = string;
// ...
export interface DataItem {
  type: string;
  metadata: JsonObject;
  payload: unknown;
}

export interface Message {
  senderRole: AgentRole;
  senderId: AgentIdentityCode;
  sessionId: string;
  taskId?: string | undefined;
  id: string;
  artifact?: MessageArtifact | undefined;
  final?: boolean | undefined;
  chunkIndex?: number | undefined;
  lastChunk?: boolean | undefined;
  dataItems: DataItem[];
  createdAt: string;
}

export interface Task {
  id: string;
  sessionId: string;
  state: TaskState;
  stateChangedAt?: string | undefined;
  messages?: Message[] | undefined;
  artifacts?: Array<Message | Task> | undefined;
}

export interface SessionReceiver {
  agentId: AgentIdentityCode;
  accessAddress?: string | undefined;
  mode: AgentInteractionMode;
  parameters?: JsonObject | undefined;
}

export interface Session {
  id: string;
  mode: AgentInteractionMode;
  sender: {
    agentId: AgentIdentityCode;
    accessAddress?: string | undefined;
  };
  receivers: SessionReceiver[];
  context?: Array<Message | Task> | undefined;
  createdAt: string;
}